praktikum demo modul 3

+++ demo Struktur Data/DemoModul1/src/main/java/org/demo1/versi1/Main.java
@@ -32,7 +32,7 @@ public class Main {
         /*Passenger<String> p = new Passenger<>("andi", "BUDI");*/
 
         Passenger<Integer> passenger = new Passenger<>(name,identity);
-        Ticket <Long> ticket = new Ticket<>(bookingCode, passenger, ticketClass);
+        Ticket <Integer> ticket = new Ticket<>(bookingCode, passenger, ticketClass);
